chore: migrate to Angular 22 and modernize codebase
- build(deps): migrate to Angular 22 + TypeScript 6.0 (ng update
  core/cli/ssr, angular-eslint 22); remove withFetch() (fetch défaut v22);
  clean migration cruft (\$safeNavigationMigration, global diagnostic
  suppression); fix unnecessary optional chain (NG8107)
- fix(lint): restore standalone ESLint flat config (dangling AAK preset)
- refactor: remove AI-style decorative comments and narrative JSDoc
- refactor(calendar): rename RelativeTimePipe -> RelativeTime
- refactor: split 5 god-components into ~11 dumb subcomponents (contact-form,
  project-detail, admin-projects, admin-analytics, admin-project-inline-form);
  move getComputedStyle out of computed() into a theme palette signal
