+++ eslint.config.js
@@ -4,12 +4,7 @@ const tseslint = require("typescript-eslint");
 const angular = require("angular-eslint");
 const prettierConfig = require("eslint-config-prettier");
 
-module.exports = (async () => {
-  // Préréglage AAK vendoré (.mjs/ESM) — importé dynamiquement car cette config
-  // est en CommonJS. ESLint 9 flat config accepte un export Promise.
-  const aak = (await import("./.claude/eslint/aak-conventions.mjs")).default;
-
-  return [
+module.exports = [
   {
     files: ["**/*.ts"],
     ...eslint.configs.recommended,
@@ -30,7 +25,6 @@ module.exports = (async () => {
     files: ["**/*.ts"],
     processor: angular.processInlineTemplates,
     rules: {
-      // Angular selectors
       "@angular-eslint/directive-selector": [
         "error",
         {
@@ -47,8 +41,6 @@ module.exports = (async () => {
           style: "kebab-case",
         },
       ],
-
-      // TypeScript strict
       "@typescript-eslint/consistent-type-definitions": ["error", "type"],
       "@typescript-eslint/no-explicit-any": "error",
       "@typescript-eslint/explicit-function-return-type": "warn",
@@ -56,8 +48,8 @@ module.exports = (async () => {
         "error",
         {
           argsIgnorePattern: "^_",
-          varsIgnorePattern: "^_"
-        }
+          varsIgnorePattern: "^_",
+        },
       ],
     },
   },
@@ -69,20 +61,6 @@ module.exports = (async () => {
     files: ["**/*.html"],
     ...config,
   })),
-  {
-    files: ["**/*.html"],
-    rules: {},
-  },
-  // Prettier DOIT être en dernier
+  // Prettier en dernier : désactive les règles stylistiques gérées par Prettier.
   prettierConfig,
-  // Préréglage AAK (no-restricted-syntax, …) — après prettier : aucune règle
-  // stylistique que prettier désactive.
-  ...aak,
-  {
-    // Point d'entrée SSR : `export default bootstrap` imposé par Angular SSR —
-    // hors périmètre de la règle AAK (réservée aux pages routées *-page.ts).
-    files: ["**/main.server.ts"],
-    rules: { "no-restricted-syntax": "off" },
-  },
-  ];
-})();
+];
