+++ package.json
@@ -48,23 +48,23 @@
   },
   "packageManager": "pnpm@10.22.0",
   "devDependencies": {
-    "@angular-eslint/builder": "21.4.0",
-    "@angular/build": "^21.2.13",
-    "@angular/cli": "^21.2.13",
-    "@angular/common": "^21.2.15",
-    "@angular/compiler": "^21.2.15",
-    "@angular/compiler-cli": "^21.2.15",
-    "@angular/core": "^21.2.15",
-    "@angular/forms": "^21.2.15",
-    "@angular/platform-browser": "^21.2.15",
-    "@angular/platform-server": "^21.2.15",
-    "@angular/router": "^21.2.15",
-    "@angular/ssr": "^21.2.13",
+    "@angular-eslint/builder": "22.0.0",
+    "@angular/build": "^22.0.4",
+    "@angular/cli": "^22.0.4",
+    "@angular/common": "^22.0.3",
+    "@angular/compiler": "^22.0.3",
+    "@angular/compiler-cli": "^22.0.3",
+    "@angular/core": "^22.0.3",
+    "@angular/forms": "^22.0.3",
+    "@angular/platform-browser": "^22.0.3",
+    "@angular/platform-server": "^22.0.3",
+    "@angular/router": "^22.0.3",
+    "@angular/ssr": "^22.0.4",
     "@eslint/js": "^10.0.1",
     "@fortawesome/fontawesome-free": "^6.7.2",
     "@sentry/cli": "^3.5.0",
     "@tailwindcss/postcss": "^4.3.0",
-    "angular-eslint": "21.4.0",
+    "angular-eslint": "22.0.0",
     "chart.js": "^4.5.1",
     "eslint": "^10.4.1",
     "eslint-config-prettier": "^10.1.8",
@@ -80,7 +80,7 @@
     "tailwindcss": "^4.3.0",
     "tslib": "^2.3.0",
     "tsx": "^4.22.4",
-    "typescript": "~5.9.2",
+    "typescript": "~6.0.3",
     "typescript-eslint": "8.60.1",
     "vitest": "^4.1.8"
   },
