+++ src/app/features/admin/application/admin-projects.ts
@@ -9,23 +9,21 @@ import {
 import { takeUntilDestroyed, rxResource } from '@angular/core/rxjs-interop';
 import { firstValueFrom, switchMap } from 'rxjs';
 import { FormsModule } from '@angular/forms';
-import { NgOptimizedImage } from '@angular/common';
 import { ProjectsGateway } from '@features/projects/domain/gateways/projects.gateway';
 import type { Project, ProjectInput } from '@features/projects/domain/models/project.model';
 import { HomeGateway } from '@features/home/domain/gateways/home.gateway';
 import { AdminProjectInlineForm } from './components/admin-project-inline-form';
-import { AppTag } from '@shared/ui/tag';
+import { AdminProjectRow } from './components/admin-project-row';
 import { ToastStore } from '@shared/ui/toast-store';
 import { Button } from '@shared/ui/button';
 import { AppIcon } from '@shared/icons/app-icon';
 
 @Component({
   selector: 'app-admin-projects',
-  imports: [AdminProjectInlineForm, FormsModule, NgOptimizedImage, AppTag, AppIcon, Button],
+  imports: [AdminProjectInlineForm, AdminProjectRow, FormsModule, AppIcon, Button],
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
     <div>
-      <!-- Header -->
       <div class="flex items-center justify-between mb-8">
         <h1 class="text-2xl font-bold text-foreground">Projets</h1>
         <div class="flex items-center gap-4">
@@ -55,7 +53,6 @@ import { AppIcon } from '@shared/icons/app-icon';
         </div>
       </div>
 
-      <!-- New project form -->
       @if (showNewForm()) {
         <div class="mb-6">
           <app-admin-project-inline-form
@@ -65,95 +62,16 @@ import { AppIcon } from '@shared/icons/app-icon';
         </div>
       }
 
-      <!-- Projects list -->
       <div class="space-y-3">
-        @for (project of filteredProjects(); track project.id) {
-          <div class="bg-surface border border-foreground/10 rounded-xl overflow-hidden">
-            <!-- Card row -->
-            <div class="flex items-center gap-4 px-5 py-4">
-              <!-- Thumbnail -->
-              <div class="shrink-0">
-                @if (project.image) {
-                  <img
-                    [ngSrc]="project.image"
-                    [alt]="project.title"
-                    width="48"
-                    height="48"
-                    class="w-12 h-12 rounded-lg object-cover"
-                  />
-                } @else {
-                  <div
-                    class="w-12 h-12 rounded-lg bg-foreground/10 flex items-center justify-center"
-                  >
-                    <svg
-                      class="w-6 h-6 text-muted"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      stroke="currentColor"
-                    >
-                      <path
-                        stroke-linecap="round"
-                        stroke-linejoin="round"
-                        stroke-width="1.5"
-                        d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z"
-                      />
-                    </svg>
-                  </div>
-                }
-              </div>
-
-              <!-- Title + tags -->
-              <div class="flex-1 min-w-0">
-                <p class="text-sm font-medium text-foreground truncate">{{ project.title }}</p>
-                <div class="flex flex-wrap gap-1 mt-1">
-                  @for (tag of project.tags.slice(0, 4); track tag) {
-                    <app-tag [value]="tag" severity="info" />
-                  }
-                  @if (project.tags.length > 4) {
-                    <app-tag [value]="'+' + (project.tags.length - 4)" severity="secondary" />
-                  }
-                </div>
-              </div>
-
-              <!-- Category + featured -->
-              <div class="hidden sm:flex flex-col items-end gap-1 shrink-0">
-                <span class="text-xs text-muted">{{ project.category }}</span>
-                @if (project.featured) {
-                  <app-tag value="Featured" severity="warn" />
-                }
-              </div>
-
-              <!-- Actions -->
-              <div class="flex items-center gap-2 shrink-0">
-                <app-button
-                  severity="secondary"
-                  variant="outlined"
-                  [ariaLabel]="editingId() === project.id ? 'Fermer' : 'Modifier'"
-                  (click)="toggleEdit(project.id)"
-                >
-                  @if (editingId() === project.id) {
-                    <app-icon name="times" [size]="20" />
-                  } @else {
-                    <app-icon name="pencil" [size]="20" />
-                  }
-                </app-button>
-                <app-button severity="danger" ariaLabel="Supprimer" (click)="deleteProject(project)">
-                  <app-icon name="trash" [size]="20" />
-                </app-button>
-              </div>
-            </div>
-
-            <!-- Inline edit form -->
-            @if (editingId() === project.id) {
-              <div class="px-5 pb-5">
-                <app-admin-project-inline-form
-                  [project]="project"
-                  (saved)="updateProject(project.id, $event)"
-                  (cancelled)="editingId.set(null)"
-                />
-              </div>
-            }
-          </div>
+        @for (proj of filteredProjects(); track proj.id) {
+          <app-admin-project-row
+            [project]="proj"
+            [isEditing]="editingId() === proj.id"
+            (editToggled)="toggleEdit(proj.id)"
+            (deleteClicked)="deleteProject(proj)"
+            (saved)="updateProject(proj.id, $event)"
+            (cancelled)="editingId.set(null)"
+          />
         } @empty {
           <div
             class="text-center py-12 text-muted text-sm bg-surface border border-foreground/10 rounded-xl"
