+++ src/app/features/admin/application/components/admin-project-form-data.ts
@@ -1,0 +1,44 @@
+export const PROJECT_CATEGORIES = [
+  'Application Web',
+  'Application Mobile',
+  'API / Backend',
+  'Script',
+  'Package / Librairie',
+  'Extension',
+  'Design / Maquette',
+] as const;
+
+export const AVAILABLE_PROJECT_TAGS = [
+  'Angular',
+  'TypeScript',
+  'JavaScript',
+  'NestJS',
+  'Node.js',
+  'Python',
+  'Bash',
+  'TailwindCSS',
+  'SCSS',
+  'CSS',
+  'PostgreSQL',
+  'MongoDB',
+  'Supabase',
+  'Redis',
+  'SQLite',
+  'Docker',
+  'CI/CD',
+  'GitHub Actions',
+  'Nginx',
+  'Linux',
+  'Git',
+  'JWT',
+  'API',
+  'REST',
+  'GraphQL',
+  'WebSocket',
+  'CLI',
+  'Automation',
+  'DevOps',
+  'Vitest',
+  'SSR',
+  'Astro',
+] as const;
